Refactor - แก้ให้ส่งข้อมูลเป็น float + ทำให้โหนดที่เพิ่มเองถูกรวมใน node_count

+++ internal/platform/aiclient/types.go
@@ -26,9 +26,9 @@ type SearchResponse struct {
 type SentimentRequest struct {
 	LearningReflect  string `json:"learning_reflect"`
 	MoodReflect      string `json:"mood_reflect"`
-	FeelScore        int    `json:"feel_score"`
-	ProgressScore    int    `json:"progress_score"`
-	ChallengeScore   int    `json:"challenge_score"`
+	FeelScore        float64 `json:"feel_score"`
+	ProgressScore    float64 `json:"progress_score"`
+	ChallengeScore   float64 `json:"challenge_score"`
 }
 
 // SentimentResponse represents the sentiment analysis response
